+++ drape/buffer_base.hpp
@@ -8,25 +8,25 @@ namespace dp
 class BufferBase
 {
 public:
-  BufferBase(uint8_t elementSize, uint16_t capacity);
+  BufferBase(uint8_t elementSize, uint32_t capacity);
 
-  uint16_t GetCapacity() const;
-  uint16_t GetCurrentSize() const;
-  uint16_t GetAvailableSize() const;
+  uint32_t GetCapacity() const;
+  uint32_t GetCurrentSize() const;
+  uint32_t GetAvailableSize() const;
 
   uint8_t GetElementSize() const;
 
-  void Seek(uint16_t elementNumber);
+  void Seek(uint32_t elementNumber);
 
 protected:
-  void Resize(uint16_t elementCount);
-  void UploadData(uint16_t elementCount);
-  void SetDataSize(uint16_t elementCount);
+  void Resize(uint32_t elementCount);
+  void UploadData(uint32_t elementCount);
+  void SetDataSize(uint32_t elementCount);
 
 private:
   uint8_t m_elementSize;
-  uint16_t m_capacity;
-  uint16_t m_size;
+  uint32_t m_capacity;
+  uint32_t m_size;
 };
 
 } // namespace dp
